Move connection events our of tool
So they can be imported without a dependency on GTK.
