+++ gaphor/diagram/connectors.py
@@ -11,10 +11,13 @@ import itertools
 from typing import Iterator, Protocol, TypeVar
 
 from gaphas.connections import Connection
+from gaphas.connector import ConnectionSink
+from gaphas.connector import Connector as ConnectorAspect
 from gaphas.connector import Handle, Port
 from generic.multidispatch import FunctionDispatcher, multidispatch
 
 from gaphor.core.modeling import Diagram, Element, Presentation
+from gaphor.core.modeling.event import RevertibeEvent
 from gaphor.core.modeling.properties import association, redefine, relation
 from gaphor.diagram.copypaste import copy, paste
 from gaphor.diagram.presentation import ElementPresentation, LinePresentation
@@ -427,3 +430,78 @@ def paste_model(copy_data, diagram) -> Iterator[Element]:
         element.postload()
 
     return iter(new_elements.values())
+
+
+class ItemConnected(RevertibeEvent):
+    def __init__(self, element, handle, connected, port):
+        super().__init__(element)
+        self.handle_index = element.handles().index(handle)
+        self.port_index = connected.ports().index(port)
+
+    def revert(self, target):
+        # Reverse only the diagram level connection.
+        # Associations have their own handlers
+        connections = target.diagram.connections
+        handle = target.handles()[self.handle_index]
+
+        connector = ConnectorAspect(target, handle, connections)
+        if cinfo := connections.get_connection(handle):
+            cinfo.callback.disable = True
+        connector.disconnect()
+
+
+class ItemDisconnected(RevertibeEvent):
+    def __init__(self, element, handle, connected, port):
+        super().__init__(element)
+        self.handle_index = element.handles().index(handle)
+        self.connected_id = connected.id
+        self.port_index = connected.ports().index(port)
+
+    def revert(self, target):
+        # Reverse only the diagram level connection.
+        # Associations have their own handlers
+        connections = target.diagram.connections
+        connected = target.diagram.lookup(self.connected_id)
+        sink = ConnectionSink(connected)
+        sink.port = connected.ports()[self.port_index]
+        handle = target.handles()[self.handle_index]
+
+        connector = ConnectorAspect(target, handle, connections)
+        connector.connect_handle(sink)
+        target.handle(ItemConnected(target, handle, sink.item, sink.port))
+
+
+class ItemTemporaryDisconnected(RevertibeEvent):
+    def __init__(self, element, handle, connected, port):
+        super().__init__(element)
+        self.handle_index = element.handles().index(handle)
+        self.connected_id = connected.id
+        self.port_index = connected.ports().index(port)
+
+    def revert(self, target):
+        connections = target.diagram.connections
+        connected = target.diagram.lookup(self.connected_id)
+        sink = ConnectionSink(connected)
+        sink.port = connected.ports()[self.port_index]
+        handle = target.handles()[self.handle_index]
+
+        connections.reconnect_item(
+            target, handle, sink.port, sink.constraint(target, handle)
+        )
+        target.handle(ItemReconnected(target, handle))
+
+
+class ItemReconnected(RevertibeEvent):
+    def __init__(self, element, handle):
+        super().__init__(element)
+        self.handle_index = element.handles().index(handle)
+
+    def revert(self, target):
+        connections = target.diagram.connections
+        handle = target.handles()[self.handle_index]
+        cinfo = connections.get_connection(handle)
+
+        connections.solver.remove_constraint(cinfo.constraint)
+        target.handle(
+            ItemTemporaryDisconnected(target, handle, cinfo.connected, cinfo.port)
+        )
