+++ gaphor/diagram/tools/connector.py
@@ -1,6 +1,5 @@
 import logging
 
-from gaphas.connector import ConnectionSink
 from gaphas.connector import Connector as ConnectorAspect
 from gaphas.connector import ItemConnector, LineConnector
 from gaphas.handlemove import HandleMove, ItemHandleMove
@@ -8,8 +7,13 @@ from gaphas.types import Pos
 
 from gaphor.core import transactional
 from gaphor.core.modeling import Presentation
-from gaphor.core.modeling.event import RevertibeEvent
-from gaphor.diagram.connectors import Connector
+from gaphor.diagram.connectors import (
+    Connector,
+    ItemConnected,
+    ItemDisconnected,
+    ItemReconnected,
+    ItemTemporaryDisconnected,
+)
 from gaphor.diagram.presentation import ElementPresentation, LinePresentation
 
 log = logging.getLogger(__name__)
@@ -115,78 +119,3 @@ class DisconnectHandle:
             adapter = Connector(connected, item)
             adapter.disconnect(handle)
         item.handle(ItemDisconnected(item, handle, connected, port))
-
-
-class ItemConnected(RevertibeEvent):
-    def __init__(self, element, handle, connected, port):
-        super().__init__(element)
-        self.handle_index = element.handles().index(handle)
-        self.port_index = connected.ports().index(port)
-
-    def revert(self, target):
-        # Reverse only the diagram level connection.
-        # Associations have their own handlers
-        connections = target.diagram.connections
-        handle = target.handles()[self.handle_index]
-
-        connector = ConnectorAspect(target, handle, connections)
-        if cinfo := connections.get_connection(handle):
-            cinfo.callback.disable = True
-        connector.disconnect()
-
-
-class ItemDisconnected(RevertibeEvent):
-    def __init__(self, element, handle, connected, port):
-        super().__init__(element)
-        self.handle_index = element.handles().index(handle)
-        self.connected_id = connected.id
-        self.port_index = connected.ports().index(port)
-
-    def revert(self, target):
-        # Reverse only the diagram level connection.
-        # Associations have their own handlers
-        connections = target.diagram.connections
-        connected = target.diagram.lookup(self.connected_id)
-        sink = ConnectionSink(connected)
-        sink.port = connected.ports()[self.port_index]
-        handle = target.handles()[self.handle_index]
-
-        connector = ConnectorAspect(target, handle, connections)
-        connector.connect_handle(sink)
-        target.handle(ItemConnected(target, handle, sink.item, sink.port))
-
-
-class ItemTemporaryDisconnected(RevertibeEvent):
-    def __init__(self, element, handle, connected, port):
-        super().__init__(element)
-        self.handle_index = element.handles().index(handle)
-        self.connected_id = connected.id
-        self.port_index = connected.ports().index(port)
-
-    def revert(self, target):
-        connections = target.diagram.connections
-        connected = target.diagram.lookup(self.connected_id)
-        sink = ConnectionSink(connected)
-        sink.port = connected.ports()[self.port_index]
-        handle = target.handles()[self.handle_index]
-
-        connections.reconnect_item(
-            target, handle, sink.port, sink.constraint(target, handle)
-        )
-        target.handle(ItemReconnected(target, handle))
-
-
-class ItemReconnected(RevertibeEvent):
-    def __init__(self, element, handle):
-        super().__init__(element)
-        self.handle_index = element.handles().index(handle)
-
-    def revert(self, target):
-        connections = target.diagram.connections
-        handle = target.handles()[self.handle_index]
-        cinfo = connections.get_connection(handle)
-
-        connections.solver.remove_constraint(cinfo.constraint)
-        target.handle(
-            ItemTemporaryDisconnected(target, handle, cinfo.connected, cinfo.port)
-        )
